ProyectoFinal/requirements.txt
streamlit
opencv-python
numpy
pillow
insightface
onnxruntime
